Remove properties from ZoomTelescope and move zoomtelescope.py into a separate lasers folder.
